The feature extraction methods (looking for similar blocks) have been preliminarily improved, but these methods need to communicate with the persistence class as a message queue, which is not yet complete

+++ include/struct.h
@@ -17,8 +17,8 @@
 typedef struct
 {
     uint64_t chunkID;
-    uint32_t chunkSize;
-    uint32_t saveSize;
+    uint64_t chunkSize;
+    uint64_t saveSize;
     uint8_t *chunkPtr;
     int basechunkid;
     uint8_t deltaFlag;
